psplash.c: fix crash from length becoming negative
If there is an error in read(), it returns -1 but this is just added to
length without checking first. This can lead to a runaway negative
value for length which eventually crashes when memchr() is called with the
negative value.

The fix is to check the return from read() first and handle the error
state.

[YOCTO #14806]

Signed-off-by: Chris Moore <cmoore@hanoverdisplays.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>