configure.ac: standardize default-enabled options
This commit makes the features progress-bar and startup-msg enabled by
default, and enabled when respectively `--enable-progress-bar` or
`--enable-startup-msg` is passed.

To do so, the default variable name `enable_feature` is used, as well as
the default variable values `yes` and `no`.

This is linked to this issue on buildroot
https://gitlab.com/buildroot.org/buildroot/-/issues/112

The documentation for this feature of autoconf is available here:
https://autotools.info/autoconf/arguments.html

Signed-off-by: Florent Allard <florent.allard@savoirfairelinux.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>